pyproject.toml
[project]
name = "reflection"
version = "0.0.0"
requires-python = ">=3.13,<3.14"
dependencies = [
  "numpy",
  "pandas",
  "pyarrow",
  "matplotlib",
  "seaborn",
  "pandas-datareader",
  "mplfinance",
  "langchain",
  "langgraph",
  "langchain-openai",
  "langchain-aws",
  "pydantic",
  "langgraph-checkpoint",
  "langgraph-checkpoint-sqlite",
  "fastapi",
  "uvicorn[standard]",
  "dependency-injector",
  "pyyaml",
  "rank-bm25",
]


[project.optional-dependencies]
dev = [
  "ruff",
  "ipykernel",
  "pytest",
  "pytest-asyncio",
  "httpx",
]
